# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Music and Album API",
      "version": "1.0.0",
      "description": "API para gerenciar músicas e álbuns",
      "contact": {
        "name": "SavanaPoint Dev Team"
      }
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ],
    "components": {
      "schemas": {
        "Audio": {
          "type": "object",
          "required": ["title", "artist", "genre", "duration", "url", "artistAvatar"],
          "properties": {
            "title": {
              "type": "string",
              "description": "Título da música"
            },
            "artist": {
              "type": "string",
              "description": "Nome do artista"
            },
            "artistAvatar": {
              "type": "string",
              "description": "URL do avatar do artista"
            },
            "album": {
              "type": "string",
              "description": "Nome do álbum"
            },
            "genre": {
              "type": "string",
              "description": "Gênero da música"
            },
            "duration": {
              "type": "number",
              "description": "Duração da música em segundos"
            },
            "releaseDate": {
              "type": "string",
              "format": "date",
              "description": "Data de lançamento da música"
            },
            "url": {
              "type": "string",
              "description": "URL do arquivo de áudio, onde o arquivo pode ser reproduzido"
            },
            "cover": {
              "type": "string",
              "description": "URL da capa da música ou álbum"
            },
            "producer": {
              "type": "string",
              "description": "Produtor da música"
            },
            "composer": {
              "type": "string",
              "description": "Compositor da música"
            },
            "category": {
              "type": "string",
              "enum": ["Rock", "Pop", "Jazz", "Classical", "Hip-hop", "Electronic", "Other"],
              "description": "Categoria da música"
            },
            "playCount": {
              "type": "number",
              "description": "Número de reproduções da música"
            }
          },
          "example": {
            "title": "My Song",
            "artist": "John Doe",
            "artistAvatar": "https://example.com/avatar.jpg",
            "album": "My Album",
            "genre": "Rock",
            "duration": 180,
            "releaseDate": "2024-01-01",
            "url": "https://example.com/audio.mp3",
            "cover": "https://example.com/cover.jpg",
            "producer": "Jane Smith",
            "composer": "John Doe",
            "category": "Rock",
            "playCount": 100
          }
        },
        "Album": {
          "type": "object",
          "required": ["title", "artist", "genre", "artistAvatar", "trackIds"],
          "properties": {
            "title": {
              "type": "string",
              "description": "Título do álbum"
            },
            "artist": {
              "type": "string",
              "description": "Nome do artista"
            },
            "artistAvatar": {
              "type": "string",
              "description": "URL do avatar do artista"
            },
            "genre": {
              "type": "string",
              "description": "Gênero do álbum"
            },
            "releaseDate": {
              "type": "string",
              "format": "date",
              "description": "Data de lançamento do álbum"
            },
            "cover": {
              "type": "string",
              "description": "URL da capa do álbum"
            },
            "trackIds": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "Lista de IDs das músicas no álbum"
            }
          },
          "example": {
            "title": "My Album",
            "artist": "John Doe",
            "artistAvatar": "https://example.com/avatar.jpg",
            "genre": "Rock",
            "releaseDate": "2024-01-01",
            "cover": "https://example.com/cover.jpg",
            "trackIds": ["64b0c34d0f28e3b8b8e87fa9", "64b0c34d0f28e3b8b8e87faa"]
          }
        }
      }
    },
    "paths": {
      "/upload": {
        "post": {
          "summary": "Faz o upload de uma música",
          "tags": ["Audio"],
          "requestBody": {
            "required": true,
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "audioFile": {
                      "type": "string",
                      "format": "binary",
                      "description": "Arquivo de áudio para upload"
                    },
                    "title": {
                      "type": "string",
                      "description": "Título da música"
                    },
                    "artist": {
                      "type": "string",
                      "description": "Nome do artista"
                    },
                    "genre": {
                      "type": "string",
                      "description": "Gênero da música"
                    },
                    "duration": {
                      "type": "number",
                      "description": "Duração da música em segundos"
                    },
                    "artistAvatar": {
                      "type": "string",
                      "description": "URL do avatar do artista"
                    },
                    "cover": {
                      "type": "string",
                      "description": "URL da capa da música"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Música enviada com sucesso",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Áudio enviado com sucesso!"
                      },
                      "audio": {
                        "$ref": "#/components/schemas/Audio"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Falha no upload"
            }
          }
        }
      },
      "/audios": {
        "get": {
          "summary": "Lista todas as músicas com paginação e filtros",
          "tags": ["Audio"],
          "parameters": [
            {
                "name": "url",
                "in": "query",
                "description": "Filtrar url do audio",
                "schema": {
                  "type": "string"
                }
              },
            {
              "name": "artist",
              "in": "query",
              "description": "Filtrar por nome do artista",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "genre",
              "in": "query",
              "description": "Filtrar por gênero",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "album",
              "in": "query",
              "description": "Filtrar por nome do álbum",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "category",
              "in": "query",
              "description": "Filtrar por categoria",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "page",
              "in": "query",
              "description": "Número da página",
              "schema": {
                "type": "integer"
              }
            },
            {
              "name": "limit",
              "in": "query",
              "description": "Limite de itens por página",
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Lista de músicas",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Audio"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/audios/{id}/play": {
        "post": {
          "summary": "Incrementa a contagem de reproduções de uma música",
          "tags": ["Audio"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID da música",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Contagem de reproduções atualizada",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Audio"
                  }
                }
              }
            },
            "404": {
              "description": "Música não encontrada"
            }
          }
        }
      },
      "/albums": {
        "post": {
          "summary": "Cria um novo álbum",
          "tags": ["Album"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Album"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Álbum criado com sucesso",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Album"
                  }
                }
              }
            },
            "400": {
              "description": "Erro na criação do álbum"
            }
          }
        },
        "get": {
          "summary": "Lista todos os álbuns",
          "tags": ["Album"],
          "parameters": [
            {
              "name": "artist",
              "in": "query",
              "description": "Filtrar por nome do artista",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Lista de álbuns",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Album"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/albums/{id}": {
        "put": {
          "summary": "Atualiza um álbum existente",
          "tags": ["Album"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID do álbum",
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Album"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Álbum atualizado com sucesso",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Album"
                  }
                }
              }
            },
            "404": {
              "description": "Álbum não encontrado"
            },
            "403": {
              "description": "Permissão negada"
            }
          }
        },
        "delete": {
          "summary": "Exclui um álbum",
          "tags": ["Album"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID do álbum",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Álbum excluído com sucesso"
            },
            "404": {
              "description": "Álbum não encontrado"
            },
            "403": {
              "description": "Permissão negada"
            }
          }
        }
      }
    }
  }
  